complete refactor to free
