--- modules/async/async-cats-effect/shared/src/main/scala/cats-effect/AsyncCatsEffect.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2017 47 Degrees, LLC. <http://www.47deg.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package freestyle.asyncCatsEffect
-
-import freestyle.free.async._
-import cats.effect.Effect
-
-trait AsyncCatsEffectImplicits {
-  implicit def catsEffectAsyncContext[F[_]](implicit F: Effect[F]): AsyncContext[F] =
-    new AsyncContext[F] {
-      def runAsync[A](fa: Proc[A]): F[A] = F.async(fa)
-    }
-}
-
-object implicits extends AsyncCatsEffectImplicits
