+++ modules/examples/todolist/src/main/scala/todo/http/apis/TodoListApi.scala
@@ -24,7 +24,7 @@ import io.finch._
 import io.finch.circe._
 import io.circe.generic.auto._
 import freestyle.free._
-import freestyle.http.finch._
+import freestyle.free.http.finch._
 import todo.model.TodoList
 import todo.services._
