--- modules/integrations/cache-redis/src/main/scala/rediscala/package.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright 2017 47 Degrees, LLC. <http://www.47deg.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package freestyle.cache.redis
-
-import cats.data.Kleisli
-import _root_.redis.commands.{Keys, Server, Strings}
-
-package object rediscala {
-
-  type Commands = Keys with Server with Strings
-
-  type Ops[F[_], A] = Kleisli[F, Commands, A]
-
-}
