+++ modules/integrations/monix/src/main/scala/free/monix.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017 47 Degrees, LLC. <http://www.47deg.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package freestyle.free
+
+import freestyle.free._
+import monix.eval.Task
+
+object monixCapture {
+
+  implicit val freestyleMonixCaptureInstance: Capture[Task] = new Capture[Task] {
+    override def capture[A](a: => A) = Task(a)
+  }
+
+}
